refactor: FocusHeader 컴포넌트 css 분리
