+++ src/pages/TodayFocusPage/TodayFocus.module.css
@@ -5,29 +5,6 @@
   font-weight: 500;
 }
 
-.header {
-  display: flex;
-  justify-content: space-between;
-  align-items: center;
-}
-
-.navContainer {
-  display: flex;
-  gap: 16px;
-}
-
-.nav {
-  display: flex;
-  align-items: center;
-  gap: 4px;
-  padding: 12px 16px 12px 24px;
-  border-radius: 15px;
-  border: 1px solid var(--gray_DDDDDD);
-  background: var(--white_FFFFFF);
-  color: var(--gray_818181);
-  font-weight: 500;
-}
-
 .pointContainer {
   display: flex;
   flex-direction: column;
@@ -37,21 +14,6 @@
 .pointContainer > p {
   font-size: 18px;
   color: var(--gray_818181);
-}
-
-.point {
-  display: flex;
-  width: fit-content;
-  align-items: center;
-  padding: 8px 12px;
-  gap: 4px;
-  border-radius: 50px;
-  border: 1px solid rgba(0, 0, 0, 0.1);
-  background: rgba(255, 255, 255, 0.3);
-}
-.point > img {
-  width: 19px;
-  height: 19px;
 }
 
 .timerWrapper {
@@ -83,24 +45,6 @@
 @media (max-width: 768px) {
   .focusWrapper {
     gap: 24px;
-  }
-
-  .header {
-    flex-direction: column;
-    align-items: start;
-    gap: 16px;
-  }
-
-  .title {
-    font-size: 24px;
-  }
-
-  .navContainer {
-    gap: 8px;
-  }
-
-  .nav {
-    padding: 8px 6px 8px 10px;
   }
 
   .timerWrapper {
